Break out formatters
In the change made to move to std::format, we defined some custom type
formatters in logging.hpp.  This had the unintended effect of making
all compile units pull in the majority of boost::url, and nlohmann::json
as includes.

This commit breaks out boost and json formatters into their own separate
includes.

Tested: Code compiles.  Logging changes only.
