+++ include/openbmc_dbus_rest.hpp
@@ -15,10 +15,12 @@
 #pragma once
 #include "app.hpp"
 #include "async_resp.hpp"
+#include "boost_formatters.hpp"
 #include "dbus_singleton.hpp"
 #include "dbus_utility.hpp"
 #include "http_request.hpp"
 #include "http_response.hpp"
+#include "json_formatters.hpp"
 #include "logging.hpp"
 #include "parsing.hpp"
 #include "routing.hpp"
